refactor: Move keyboard monitor to `sys.ui.dom`
