+++ code/system.ui/dom/src/common/libs.mts
@@ -1,9 +1,12 @@
 /**
  * @external
  */
+import { clone, uniq } from 'ramda';
+export const R = { clone, uniq };
+
 export { Subject } from 'rxjs';
 
 /**
  * @system
  */
-export { Time, Path } from 'sys.util';
+export { Time, Path, slug, rx } from 'sys.util';
