+++ code/system.ui/dom/src/index.mts
@@ -2,5 +2,6 @@ export * from './types.mjs';
 export { Pkg } from './index.pkg.mjs';
 
 export { KeyListener } from './web.events';
+export { Keyboard, KeyboardMonitor } from './web.Keyboard';
 export { LocalStorage } from './web.LocalStorage';
-export { WrangleUrl } from './web.url';
+export { WrangleUrl } from './web.Url';
