+++ code/system.ui/dom/src/types.mts
@@ -1,1 +1,3 @@
+export * from './web.events/types.mjs';
+export * from './web.Keyboard/types.mjs';
 export * from './web.LocalStorage/types.mjs';
