+++ code/system.ui/react.common/src/common/types.mts
@@ -15,8 +15,10 @@ export type {
   EventBus,
   Disposable,
 } from 'sys.types/src/types.mjs';
+
 export type { CssValue, CssEdgesInput, CssShadow } from 'sys.ui.react.css/src/types.mjs';
 export type { TestSuiteRunResponse, TestRunResponse } from 'sys.test.spec/src/types.mjs';
+
 export type {
   DevCtx,
   DevCtxInput,
@@ -32,6 +34,16 @@ export type {
   DevRenderer,
   DevRenderProps,
 } from 'sys.ui.react.dev.types/src/types.mjs';
+
+export type {
+  UIEventBase,
+  UIModifierKeys,
+  KeyboardState,
+  KeyboardKeypress,
+  KeyboardKeypressProps,
+  KeyboardModifierEdges,
+  KeyboardModifierFlags,
+} from 'sys.ui.dom/src/types.mjs';
 
 /**
  * @local
