+++ code/system.ui/react.common/src/ui/Event/types.mts
@@ -1,14 +1,1 @@
-export type UIEventBase = {
-  readonly bubbles: boolean;
-  readonly cancelable: boolean;
-  readonly eventPhase: number;
-  readonly timeStamp: number;
-  readonly isTrusted: boolean;
-};
-
-export type UIModifierKeys = {
-  readonly altKey: boolean;
-  readonly ctrlKey: boolean;
-  readonly metaKey: boolean;
-  readonly shiftKey: boolean;
-};
+export {};
