+++ code/system.ui/react.common/src/ui/Text.Keyboard/index.ts
@@ -1,2 +1,1 @@
 export * from './Keyboard.mjs';
-export * from './KeyboardMonitor.mjs';
