+++ code/system.ui/react.common/src/ui/types.mts
@@ -15,8 +15,6 @@ export * from './useSizeObserver/types.mjs';
 
 export * from './Text/types.mjs';
 export * from './Text.Keyboard/types.mjs';
-
-// export * from './Text.Copy/types.mjs';
 export * from './Text.Input/types.mjs';
 export * from './Text.Secret/types.mjs';
 export * from './Text.Syntax/types.mjs';
