+++ code/system.ui/react.ui.video/src/ui/common.t.ts
@@ -1,2 +1,2 @@
-import * as t from '../common/types.mjs';
+import type * as t from '../common/types.mjs';
 export { t };
